fix(charts): fix chatops-lark deployment manifest templates (#236)
This pull request includes changes to the Helm chart for the
`chatops-lark` application to update the version, simplify the NOTES.txt
template, and fix the volume mounts and volumes configuration in the
deployment template.

Version update:

*
[`charts/chatops-lark/Chart.yaml`](diffhunk://#diff-e6df411db3462581ecff43a0fa950e5dbbefb5d40548ccdca1b53db00dc415cbL18-R18):
Incremented the chart version from 0.1.0 to 0.1.1.

Template simplification:

*
[`charts/chatops-lark/templates/NOTES.txt`](diffhunk://#diff-c6cbde6d041d805f29e61d924013c3ebfd7ae81d7adf747b6760aab0bedf3b35L2-R3):
Removed the conditional block for ingress and simplified the command to
get the application URL when service type is `NodePort`.

Configuration fixes:

*
[`charts/chatops-lark/templates/deployment.yaml`](diffhunk://#diff-216f98d8d70595e33393cf796083bdd6891cd8b9f8ed053bf37cf60022551ddbL81-R92):
Corrected the indentation and usage of `volumeMounts` and `volumes`
blocks to ensure proper configuration of secrets and volumes.

Signed-off-by: wuhuizuo <wuhuizuo@126.com>